admin/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = ""
authors = [
    {name = "Lautaro Budini",email = "lautarobudini5@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy-lite (>=0.1.0,<0.2.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "geoalchemy2 (>=0.18.0,<0.19.0)"
]

[tool.poetry]
packages = [{include = "web", from = "src"}]


[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
